add forbidden delete

+++ application/controllers/Categories.php
@@ -112,6 +112,13 @@ class Categories extends CI_Controller {
 	public function delete()
 	{
 		$id = $this->input->post('category_id');
+		$material = $this->Category_model->get_material($id)->row();
+
+		if( $material != null ) {
+			$this->session->set_flashdata('used', 'Data sedang digunakan di daftar bahan, silahkan hapus semua data bahan yang berhubungan dengan kategori ini.');
+			redirect('categories');
+		}
+
 		$this->Category_model->delete($id);
 
 		$this->session->set_flashdata('deleted', 'Data berhasil dihapus.');
